Compose CanConsume and Consume to obey contract of Consumer.
Before if c := Compose(a, b) and c.CanConsume() returns true it was still
possible for c.Consume() to fail if caller fed values to a or b directly
at any time. This change fixes that. Now if c.CanConsume() returns true,
c.Consume() will succeed so long as caller doesn't feed values to a or b
between calls to c.CanConsume() and c.Consume()